Update to the 2011-06-22 release
Release notes from the web page:

* Now also available for Windows 64bit
* Easier installation (no more editing of configuration files, no
  unpacking of zip-archives)
* Automatic Adjustment of Brightness/Contrast when loading 16bit and
  float data sets
* Updated to latest HDF5 Java Library (V2.7 from February,2011)

Note that this update has one unfortunate change: three of the six
source files needed to be decompiled from bytecode, since the official
release archive did not include the sources for them. The good news is
that if someone had time, they could reconstruct the original source
from the versions present in the previous commit. But first you would
need to unify the whitespace -- then a comparison would become possible.
It is probably not worth it, though, because in two commits from now,
the entire plugin is rewritten and these sources do not survive.